backend/requirements.txt
# Backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
google-generativeai==0.3.2
openai==1.12.0
